requirements.txt
numpy==1.22.3
tensorflow-cpu==2.8.0
joblib
dill
flask
gunicorn
pandas